requirements.txt
# AI Assistant hub Backend Dependencies
# Generated from pyproject.toml for compatibility

# Web Framework
fastapi>=0.118.0
uvicorn[standard]>=0.37.0
starlette>=0.48.0

# Database
sqlalchemy>=2.0.0

# AI/ML Libraries
langchain>=0.3.27
langchain-community>=0.3.31
langchain-core>=0.3.79
langchain-huggingface>=0.3.1
langchain-ollama>=0.3.0
langchain-chroma>=0.2.6
langchain-text-splitters>=0.3.11
grandalf>=0.8

# Embeddings and Vector Store
sentence-transformers>=5.0.0
faiss-cpu>=1.12.0
huggingface-hub>=0.35.0

# text pretty print
rich>=14.2.0
unstructured>=0.6.0
markdown-it-py>=2.2.0
markdown>=2.4.2
nltk>=3.8.0

# Web Scraping
playwright>=1.55.0
beautifulsoup4>=4.14.0

# HTTP and Networking
httpx>=0.28.0
requests>=2.32.0
aiohttp>=3.13.0

# Data Processing
numpy>=2.3.0
pandas>=2.0.0
scikit-learn>=1.7.0
atlassian-python-api>=4.0.7

# Utilities
python-dotenv>=1.1.0
pydantic[email]>=2.12.0
pydantic-settings>=2.11.0
click>=8.3.0
tqdm>=4.67.0
tenacity>=9.1.0
GitPython>=3.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
authlib>=1.3.0
email-validator>=2.0.0


# Optional: Ollama for local LLM
ollama>=0.6.0

